+++ lib/git_hub_installer.dart
@@ -1,0 +1,7 @@
+library binary_repositories.git_hub_installer;
+
+import "dart:async";
+import "dart:io";
+import "package:binary_repositories/binary_repositories.dart";
+
+part "src/git_hub_installer/git_hub_installer.dart";